requirements.txt
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.56.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
keyboard==0.13.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
narwhals==1.26.0
nbformat==5.10.4
nest_asyncio==1.6.0
numpy==2.2.2
open3d==0.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
RPi.GPIO==0.7.1
rplidar==0.9.2
rplidar-roboticia==0.9.5
scipy==1.15.1
six==1.17.0
tabulate==0.9.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
